Delete Account in Auth and Firestore OK

+++ lib/Services/auth.dart
@@ -76,24 +76,26 @@ class Auth with ChangeNotifier {
     return _status;
   }
 
-    Future<AuthStatus> deleteAccount({
+  Future<AuthStatus> deleteAccount({
     required UserM user,
     required String password,
   }) async {
     try {
-      
       User newUser = _auth.currentUser!;
       user.uid = newUser.uid;
+//Reauthentication of the user
+      _status = await reAuthenticate(oldPwd: password);
 
-      //delete a Document of user into Users Collection
-     _status= await DatabaseService(uid: user.uid).deleteUser(user);
-    
-    if(_status == AuthStatus.dataDeleted){
-      //Delete User into Firebase Auth
-      await _auth.currentUser!.delete();
-    }
+      //Can delete a Document into Users Collection only if pwd entered is right
+      if (_status == AuthStatus.reauth) {
+        _status = await DatabaseService(uid: user.uid).deleteUser(user);
+      }
 
-      _status = AuthStatus.successful;
+      if (_status == AuthStatus.dataDeleted) {
+        //Delete User into Firebase Auth
+        await _auth.currentUser!.delete();
+        _status = AuthStatus.successful;
+      }
     } on FirebaseAuthException catch (e) {
       _status = AuthExceptionHandler.handleAuthException(e);
     }
